refactor(solvers): move pacakge to module.
Refactored pykp.solvers pacakage to pykp.solvers module. This
makes auto generated documentation appear in a more sensible way.
Under the previous approach, documentation for functions inside
modules in the pykp.solvers package would not render, because these
functions shared the same name as the module. Including all
functions inside a single solvers.py module rectifies this problem.
