+++ pykp/__init__.py
@@ -8,6 +8,7 @@ from .item import Item
 from .knapsack import Knapsack
 from .metrics import phase_transition, sahni_k
 from .sampler import Sampler
+from .solvers import branch_and_bound, greedy, mzn_gecode
 
 __all__ = [
     "Arrangement",
@@ -16,4 +17,7 @@ __all__ = [
     "Sampler",
     "phase_transition",
     "sahni_k",
+    "branch_and_bound",
+    "greedy",
+    "mzn_gecode",
 ]
