--- pykp/solvers/__init__.py
@@ -1,10 +1,0 @@
-"""
-This module contains implementations of various solvers for the knapsack
-problem.
-"""
-
-from .branch_and_bound import branch_and_bound
-from .greedy import greedy
-from .mzn_gecode import mzn_gecode
-
-__all__ = ["branch_and_bound", "greedy", "mzn_gecode"]
